backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-jwt-extended==4.5.2
pymongo==4.3.3
python-dotenv==1.0.0
werkzeug==2.3.4
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
gunicorn==20.1.0
